TF Data and Figure Plots
Added two folder. One folder contains all the training and validation accuracy data for the the differing networks each of which had a different varying feature. The other folder just contains the figures of the plots that have the training and validation accuracy vs the varying network features/parameter such as the number of layers and number of nodes per layer.