NO The TF Figures ARE Correct Now
Ok so fucking vscode is dumb as fuck. The matpoltlib.pyplot function to save figure savefig() is all fucked up, in general or just on my computer, and when executed in VSCode the plots don't show in the figure. So I have resorted to just manually saving the figure from with in VScode. It works now (98% sure)